GCC/Clang: Use sysroot from build system if kit doesn't set sysroot
If the kit has a sysroot set, we use that when determining the system
header paths of the tool chain for the code model.
But if the kit doesn't have a sysroot set, we should use the sysroot
setting that we find in the CXX flags that we get from the build system.

In the specific issue, we detect the Clang compiler from the Xcode
installation on macOS, which actually needs to be passed a sysroot to
return sensible system header paths.

Task-number: QTCREATORBUG-18633
Change-Id: Ida401bee8c4b82bb4fa2e6f952b8cc174ea081c6
Reviewed-by: Vikas Pachdha <vikas.pachdha@qt.io>