Wizards: introduce ObjectToFieldWidgetConverter
With it as a wrapper we can register any QObject
to use widget based field mechanism from QWizard.

It also helps to avoid the necessity for inheritance
to just adjust the text property of a widget.

Can not remove the TextCheckBox, because it is used
in some old C++ wizards.

Change-Id: I85a85a834714a4b38b501c13357fa8c8bb02b5bd
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>