Improve horizontal scrolling in the navigation views
Because of QTBUG-3927 the tree views are scrolled completely to the left
when selecting any item because it makes the header column visible. That
is very annoying, because it makes items deep in the hierarchy hardly
readable, even if you manually scrolled horizontally to see them.
Scroll horizontally to make the actual item visible instead.

Task-number: QTCREATORBUG-19204
Change-Id: Ic083236feae8892d210b9d2b1238f3c6b80a8cea
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>