Fix file system view layout on macOS
Combo boxes on Qt/macOS add an implicit spacing to their containing
widget, which lead to a gray border around the tree view.
Move the combo box into a separate widget to avoid the border around the
tree view, and paint the background of the file system view to avoid a
gray border around the combo box.

Change-Id: I9cb8329f585d48016896d558e817c3d978a78669
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>