Fix selection when switching root in file system view
Don't jump to the (previously) selected file.
If the (previously) selected file is not a child of the new root,
then "select" the root directory, which also adapts the bread crumb.
Also show the selection instead of the parent in the bread crumb,
which is more natural when a root is selected and not wrong otherwise
either.

Change-Id: Ib5a58a034101e4e2617b4ae1a5ec9188d1a92eb5
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>