Fix that double-click in file system view sometimes does not work
When clicking on an item that changes the height of the crumble path, a
double-click has great chances of not succeeding, because before the
second click, the item might move away from under the mouse.

If the tree has a scroll bar that can move wide enough, we solve that by
scrolling the tree such that the same item remains under cursor when the
crumble path height changes. (We have to synchronize the scroll bar
value change with the relayouting though, to avoid flicker.)

If there is no scroll bar, or it cannot move enough in the needed
direction, we delay the re-layouting by the maximum double-click
interval to guarantee a double-click will still have the same item under
the mouse.

Change-Id: I3b296925d9be2d2ab5affbbb64df67173d9715d4
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>