Remove qmlsceneCommand() and qmlViewerCommand() from BaseQtVersion
qmlviewer is QtQuick1 and not supported anymore.

The qmlscene command is in general not a property of the Qt version but
of the target device, as we generally assume Qt binaries to be
preinstalled on the target and qmlscene is executed on the target.

DesktopQtVersion retains a method to retrieve the qmlscene path as there
target == host and having the method available enables us to warn if the
binary is missing.

Also, QmlProjectManager needs to pick the qmlscene command from the Qt
version if it's a desktop Qt.

Change-Id: I2ee0f993e60b96d672080db2db0a6abe5ca933e3
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>