Utils: Either drop files or values
Dropping something on QtCreator will now signal either file drop or
other value drops. With the old implementation drop receiver might
not be able to differ between the two.

Also dropping files signals the drop position (which may be ignored
by the receiver).

This change is needed for a future change to fix issue
QTCREATORBUG-19117.

Change-Id: I56a72ba3d925de19979f3f90d161adc4f4986e20
Reviewed-by: Eike Ziller <eike.ziller@qt.io>