Clang: Make ClangPathWatcherNotifier non copy-able
There is already a destructor because the vtable should not be
injected in every translation unit. Because of the rule of five
we have to provide or omit the copy and move functions. So
we delete them because we don't need them and any way
it would be not smart to have to notifier.

Change-Id: I35c00954eaa4e1bb299581a5e4ba5369d5cbf1cd
Reviewed-by: Ivan Donchevskii <ivan.donchevskii@qt.io>