BaseFileFilter: Rate full matches higher than fuzzy matches
Searching for "qmap.h" should not have "qcore_mac_p.h" as
first search result and "qmap.h" somewhere later in the list.

This is only a partial solution for QTCREATORBUG-19531, as the
filter might be called multiple times (for "All Included Files"
and "Files in Any Project"). So the final result list may contain
the best results in the middle.

Task-number: QTCREATORBUG-19531
Change-Id: Ib5322824d3e0e8106c2f197169342f18923a894a
Reviewed-by: André Hartmann <aha_1980@gmx.de>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>