QmlDesigner: Fix StatesEditorModel::removeState()
We cannot simply assume that stateIndex + 1 is a valid index.
The removed state might be the last state.

This triggers an assert in Qt 5.11.

Unfortunately there seems to be an issue that forces us
to reset the model. Without reset model the ListView delivers
the wrong data to delegates, after the item was deleted.

Some minor code clean up.

Change-Id: I935e3358382371d31f0ff44c0ded229d35950648
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>