QmlDesigner: Don't blur the form editor items and background pattern
When zooming in, or generally in dpi scaled mode, it is not desirable to
have blurred items or blurred edges on the background pattern. Design
tools don't do that.

If an item is painted smaller than its natural size or if it is rotated,
it should be painted with interpolation.

Change-Id: Ie221db47a8f9a807c99ba8e0c96a2cc8d5115d6f
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>