Timeline: Improve performance of labels
a, We don't need our special synchronous reloader anymore because the
stock Loader does the same now.

b, Creating and removing the labels while scrolling only accounts for a
relatively small part of the total overhead. The most expensive part is
the offset calculation which has to be performed for each label on each
scroll movement. Move parts of that to the parent item.

c, Asynchronous loading of items in a repeater doesn't work because the
repeater will still wait until all the items are loaded. Thus, the
previous code would still load all the labels immediately when
expanding. Use the "active" property instead to load labels on demand.

d, The repeater initializes all the labels to y == 0 before moving them
into place. Make sure they aren't all initially active because of that.

Change-Id: I4d8b762899c4aa88173dacec668341f592ffe38c
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>